tests: fix failing unit tests
