+++ tests/integration/adapters/__init__.py
